move displayResult in [analytic_name].tsx to AnalyticResult to make it reusable
