+++ frontend/app/components/modules/AnalyticsResult/analytics_result.json
