--- frontend/pages/analytics/[analytic_name].module.scss
@@ -1,22 +1,0 @@
-.resultsContainer {
-  display: flex;
-  box-shadow: rgba(0, 0, 0, 0.02) 0px 1px 3px 0px,
-    rgba(27, 31, 35, 0.15) 0px 0px 0px 1px;
-  padding: 1rem;
-  .col {
-    width: 50%;
-    .field {
-      padding: 4px;
-      margin-right: 12px;
-      .label {
-        font-weight: 600;
-        margin: 0 0 8px 0;
-        font-size: 0.9rem;
-      }
-      .result {
-        margin: 0;
-        font-size: 1.1rem;
-      }
-    }
-  }
-}
